tests/submission/Microsoft/semantic/hgroup.html

<!DOCTYPE html>
<html><head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type">
  <title>verify HGROUP is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 <link href="http://www.w3.org/TR/html5/the-hgroup-element.html#the-hgroup-element" rel="help">
</head>
<body>
<div id="log"></div>
<hgroup id="hgroup">
 <h1>This is some sample text</h1>
</hgroup>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("hgroup").toString() )
 })
</script>


</body></html>
